fix: Code Cleaned Up
