+++ apps/web/components/bookings/BookingCard.tsx
@@ -1,4 +1,3 @@
-// components/bookings/BookingCard.tsx
 "use client";
 
 import React from "react";
