+++ apps/web/components/bookings/PaymentModal.tsx
@@ -1,64 +1,8 @@
-// "use client";
-
-// import React, { useState } from "react";
-// import {
-//   Dialog,
-//   DialogContent,
-//   DialogHeader,
-//   DialogTitle,
-// } from "@sambha/ui/dialog";
-// import { Button } from "@sambha/ui/button";
-// import { Input } from "@sambha/ui/input";
-
-// interface PaymentModalProps {
-//   isOpen: boolean;
-//   onClose: () => void;
-//   onPay: (amount: number) => void;
-// }
-
-// const PaymentModal: React.FC<PaymentModalProps> = ({
-//   isOpen,
-//   onClose,
-//   onPay,
-// }) => {
-//   const [amount, setAmount] = useState("");
-
-//   const handlePay = () => {
-//     if (!amount) return;
-//     onPay(Number(amount));
-//     onClose();
-//   };
-
-//   return (
-//     <Dialog open={isOpen} onOpenChange={onClose}>
-//       <DialogContent className="max-w-md md:max-w-lg">
-//         <DialogHeader>
-//           <DialogTitle>Make Payment</DialogTitle>
-//         </DialogHeader>
-//         <div className="flex flex-col gap-4">
-//           <Input
-//             type="number"
-//             placeholder="Enter amount ($)"
-//             value={amount}
-//             onChange={(e) => setAmount(e.target.value)}
-//           />
-//           <Button onClick={handlePay} className="bg-[#C96FFF] text-white">
-//             Pay Now
-//           </Button>
-//         </div>
-//       </DialogContent>
-//     </Dialog>
-
-//   );
-// };
-
-// export default PaymentModal;
 import React from "react";
 import { FaMoneyBill } from "react-icons/fa6";
 import { Notebook } from "lucide-react";
 import Image from "next/image";
 import date2 from "../../assets/images/date-2.png";
-// import { Dialog } from "@sambha/ui/dialog";
 import { Modal } from "components/Modal";
 
 const PaymentModal: React.FC = () => {
